+++ modules/harvest/modules/harvest_dashboard/harvest_dashboard.routing.yml
@@ -8,14 +8,3 @@ dkan.harvest.dashboard:
     _permission: 'dkan.haverst.dashboard'
   options:
     _auth: ['basic_auth', 'cookie']
-
-dkan.harvest.dashboard.datasets:
-  path: '/admin/dkan/harvest_dashboard/{harvestId}'
-  methods: [ GET ]
-  defaults:
-    _controller: \Drupal\harvest_dashboard\Controller\Controller::harvestDatasets
-    _title_callback: \Drupal\harvest_dashboard\Controller\Controller::harvestDatasetsTitle
-  requirements:
-    _permission: 'dkan.harvest.dashboard'
-  options:
-    _auth: [ 'basic_auth', 'cookie' ]
